strat: handle timeouts and order "failure", for alternative routes
This will allow to choose an alternate strategy if opponent is detected
during a long time.

goto functions have been renamed ('_wait' suffix) and have an extra
parameter 'wait_ms'. If the opponent is still detected after the
provided timeout, the function return.

Aliases with the previous name and signature, and using a default
timeout (for now, 2s) are provided.

A 'wait_detected_opponent()' method is provided to handle timeout on
opponent detection.